Including all entity and correct relationship between.

+++ williMeter/src/main/java/br/com/willi/williMeter/models/Product.java
@@ -18,6 +18,6 @@ public class Product {
 	private long id;
 	private String name;
 	private float price;
-	private int quantity;
+	private float quantity;
 
 }
